requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.3
django-extensions==3.1.5
django-filter==22.1
django-widget-tweaks==1.4.12
psycopg2-binary==2.9.6
sqlparse==0.4.4
typing-extensions==4.7.1
